Add conftest.py for pytest
